* MonoTorrent/MonoTorrent.Client/Managers/ConnectionManager.cs:   Accumulate returns a long and we were trying to cast a boxed long to   an int which caused an exception. Cast the return value to 'int'   before boxing it.
svn path=/trunk/bitsharp/; revision=152771

+++ src/MonoTorrent/MonoTorrent.Client/Managers/ConnectionManager.cs
@@ -106,7 +106,7 @@ namespace MonoTorrent.Client
             get
             {
                 return (int) ClientEngine.MainLoop.QueueWait (() =>
-                    Toolbox.Accumulate<TorrentManager>(torrents, (m) =>
+                    (int) Toolbox.Accumulate<TorrentManager>(torrents, (m) =>
                         m.Peers.ConnectedPeers.Count
                     )
                 );
